requirements.txt
discord.py
discord
python_dotenv
yt_dlp
asyncio 
pynacl